+++ tests/clean.sh
@@ -1,0 +1,1 @@
+rm -rf plt* nplt* dplt* Backtrace* *.out.0
